refactor: unify logging via trading logger
